- Fix some errors on SQL tables - Update SQL requests in PHP

+++ sql/Notes.sql
@@ -1,5 +1,5 @@
 CREATE TABLE Notes (
-    rating FLOAT(0, 5),
+    rating FLOAT(2, 1),
     voterID INT NOT NULL,
     rankedID INT NOT NULL,
     FOREIGN KEY (voterID) REFERENCES users(id),
